[fixup] Moved all casting rules to casting package and made them package-private. Fixed imports in ScalarOperatorGens
